move fuse under unixfs
